compliance_frontend/requirements.txt

streamlit==1.31.0
requests==2.31.0
pandas==2.1.4
plotly==5.17.0
python-dotenv==1.0.0
openpyxl==3.1.2
PyPDF2==3.0.1
reportlab==4.0.8
io==3.3.0
base64==1.0.0
json==2.0.9
datetime==5.4
time==1.0.0
uuid==1.30
